WebFinalBack/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-cleanup==8.0.0
django-cors-headers==4.3.0
django-filter==23.5.0
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
jdatetime==4.1.1
python-dateutil==2.8.2
requests==2.31.0
pillow==10.2.0
drf-nested-routers==0.93.5